File renaming to respect harness nature.
